Move tstos to utils.py
pbench/server/__init__.py has grown very complicated
for its original use, so move out tstos into utils.py.
